# SPDocsAPI/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SPDocs API",
    "version": "v1"
  },
  "paths": {
...
    "/api/Lessons/GetLessonID": {
      "get": {
        "tags": [
          "Lessons"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/Lessons/GetMultipleLessonIDs": {
      "get": {
        "tags": [
          "Lessons"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "categories",
            "type": "array",
            "items": {
              "type": "string"
            },
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/api/Lessons/GetBulkLessonIDs": {
      "get": {
        "tags": [
          "Lessons"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "category",
            "type": "string"
          },
          {
            "in": "query",
            "name": "count",
            "type": "integer",
            "format": "int32",
            "default": 1
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
...
  },
...
}